Rename IOmodule to io_module
