+++ eniric/obsolete/nIR_precision.py
@@ -11,7 +11,7 @@ import numpy as np
 from matplotlib import rc
 
 import eniric
-import eniric.iomodule as io
+import eniric.io_module as io
 import eniric.legacy
 import eniric.obsolete.plotting_functions as plt_functions
 import eniric.obsolete.snr_norm
