+++ tests/test_read_write.py
@@ -4,8 +4,7 @@ import os
 import numpy as np
 import pytest
 
-import eniric.iomodule as io
-import eniric.obsolete.IOmodule
+import eniric.io_module as io
 import eniric.utilities as utils
